+++ src/Element.php
@@ -235,16 +235,27 @@ final class Element
     {
         $selfClosing = $this->selfClosing;
 
-        /** @psalm-suppress ImpureMethodCall */
-        $writer->startElement($this->name->toString());
-        /** @psalm-suppress ImpureMethodCall */
+        /**
+         * @psalm-suppress ImpureFunctionCall
+         * @psalm-suppress PossiblyNullFunctionCall
+         */
+        (\Closure::bind(
+            fn() => $this->render($writer),
+            $this->name,
+            $this->name::class,
+        ))();
+        /**
+         * @psalm-suppress ImpureFunctionCall
+         * @psalm-suppress PossiblyNullFunctionCall
+         */
         $_ = $this
             ->attributes
             ->values()
-            ->foreach(static fn($attribute) => $writer->writeAttribute(
-                $attribute->name(),
-                $attribute->value(),
-            ));
+            ->foreach(static fn($attribute): mixed => (\Closure::bind(
+                fn() => $this->render($writer),
+                $attribute,
+                $attribute::class,
+            ))());
 
         /** @psalm-suppress ImpureMethodCall */
         $opening = Sequence::of(Str::of($writer->outputMemory()));