+++ src/Element/Name.php
@@ -11,9 +11,11 @@ use Innmind\Immutable\Maybe;
 final class Name
 {
     /**
+     * @param ?non-empty-string $namespace
      * @param non-empty-string $value
      */
     private function __construct(
+        private ?string $namespace,
         private string $value,
     ) {
     }
@@ -25,7 +27,18 @@ final class Name
      */
     public static function of(string $value): self
     {
-        return new self($value);
+        return new self(null, $value);
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @param non-empty-string $namespace
+     * @param non-empty-string $value
+     */
+    public static function namespaced(string $namespace, string $value): self
+    {
+        return new self($namespace, $value);
     }
 
     /**
@@ -38,7 +51,7 @@ final class Name
         /** @var Maybe<self> */
         return match ($value) {
             '' => Maybe::nothing(),
-            default => Maybe::just(new self($value)),
+            default => Maybe::just(new self(null, $value)),
         };
     }
 
@@ -47,6 +60,29 @@ final class Name
      */
     public function toString(): string
     {
+        if ($this->namespace !== null) {
+            return \sprintf(
+                '%s:%s',
+                $this->namespace,
+                $this->value,
+            );
+        }
+
         return $this->value;
+    }
+
+    private function render(\XMLWriter $writer): void
+    {
+        if (\is_null($this->namespace)) {
+            /** @psalm-suppress ImpureMethodCall */
+            $writer->startElement($this->value);
+        } else {
+            /** @psalm-suppress ImpureMethodCall */
+            $writer->startElementNS(
+                $this->namespace,
+                $this->value,
+                null,
+            );
+        }
     }
 }