+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Innmind\Xml\Element\Name`
 - Support for the new PHP `8.4` `\Dom\*` API
 - `Innmind\Xml\Format`
+- `Innmind\Xml\Attribute::namespaced()`
 
 ### Changed